requirements.txt
fastapi
uvicorn
pytest
httpx
SQLAlchemy
bcrypt==3.2.0
python-jose[cryptography]
